created HeaderInvoiceView component
