+++ src/components/HeaderInvoiceView.js
@@ -1,0 +1,88 @@
+import { useSelector, useDispatch } from "react-redux";
+import { useParams, useNavigate } from "react-router-dom";
+import { editInvoice, deleteInvoice } from "../store/invoicesActions";
+import { selectInvoiceById } from "../store/invoicesSlice";
+import Label from "../UI/Label";
+import { Button } from "../UI/Button";
+
+const HeaderInvoiceView = () => {
+  const { invoiceId } = useParams();
+  const dispatch = useDispatch();
+  const invoice = useSelector((state) => selectInvoiceById(state, invoiceId));
+  const navigate = useNavigate();
+
+  const handleChangeStatusInvoice = () => {
+    const updatedStatus = invoice.status === "paid" ? "pending" : "paid";
+    const updatedButtonText =
+      invoice.status === "paid" ? "Mark as Paid" : "Mark as Unpaid";
+    const updatedInvoice = { ...invoice, status: updatedStatus };
+    dispatch(editInvoice(updatedInvoice));
+    const markAsPaidButton = document.getElementById("markAsPaidButton");
+    if (markAsPaidButton) {
+      markAsPaidButton.textContent = updatedButtonText;
+    }
+  };
+
+  const handleEditInvoice = async () => {
+    const updatedInvoice = { ...invoice, status: "paid" };
+    dispatch(editInvoice(updatedInvoice));
+  };
+
+  const handleDeleteInvoice = async () => {
+    if (invoice) {
+      dispatch(deleteInvoice(invoiceId));
+      navigate("/");
+    }
+  };
+
+  return (
+    <>
+      <div className='hidden md:flex justify-between bg-light-100 dark:bg-dark-200 w-full items-center px-10 py-5 rounded-lg'>
+        <div className='md:flex items-center'>
+          <span className='text-neutral-500 text-base/2 md:mr-5'>Status </span>
+          <Label status={invoice.status} />
+        </div>
+        <div>
+          <Button
+            className='bg-neutral-100 text-neutral-500 dark:bg-dark-100 dark:text-white rounded-full px-7 py-4 mr-3'
+            title='Edit'
+            onClick={handleEditInvoice}
+          />
+          <Button
+            className='bg-red-500 text-white rounded-full px-7 py-4 mr-3'
+            title='Delete'
+            onClick={handleDeleteInvoice}
+          />
+          <Button
+            className='bg-primary-200 text-white rounded-full px-7 py-4'
+            id='markAsPaidButton'
+            title={
+              invoice.status === "paid" ? "Mark as Unpaid" : "Mark as Paid"
+            }
+            onClick={handleChangeStatusInvoice}
+          />
+        </div>
+      </div>
+      <div className='flex items-center justify-between fixed bottom-0 bg-light-100 dark:bg-dark-200 w-full p-5 md:hidden'>
+        <Button
+          className='bg-neutral-100 text-neutral-500 dark:bg-dark-100 dark:text-white rounded-full px-7 py-4 '
+          title='Edit'
+          onClick={handleEditInvoice}
+        />
+        <Button
+          className='bg-red-500 text-white rounded-full px-7 py-4'
+          title='Delete'
+          onClick={handleDeleteInvoice}
+        />
+        <Button
+          className='bg-primary-200 text-white rounded-full px-4 py-4'
+          id='markAsPaidButton'
+          title={invoice.status === "paid" ? "Mark as Unpaid" : "Mark as Paid"}
+          onClick={handleChangeStatusInvoice}
+        />
+      </div>
+    </>
+  );
+};
+
+export default HeaderInvoiceView;
